feat(api): rename default queries and variables
