+++ packages/api-client/src/api/getCategory/index.ts
@@ -1,17 +1,16 @@
 import gql from 'graphql-tag';
-import defaultQuery from './defaultQuery';
+import collectionsQuery from './collectionsQuery';
 import { ApolloQueryResult } from 'apollo-client';
 import { CustomQuery } from '@vue-storefront/core';
 import type { CategoryData, CategoryParams, Context } from '../../types';
 
 const getCategory = async (context: Context, params: CategoryParams, customQuery?: CustomQuery): Promise<ApolloQueryResult<CategoryData>> => {
-
-  const defaultVariables = {
+  const collectionsVariables = {
     ...params
   };
 
   const { collections } = context.extendQuery(customQuery,
-    { collections: { query: defaultQuery, variables: defaultVariables } }
+    { collections: { query: collectionsQuery, variables: collectionsVariables } }
   );
 
   const request = await context.client.query<CategoryData>({
