+++ packages/api-client/src/api/getFacet/index.ts
@@ -1,17 +1,17 @@
 import gql from 'graphql-tag';
-import defaultQuery from './defaultQuery';
+import searchQuery from './searchQuery';
 import { ApolloQueryResult } from 'apollo-client';
 import { CustomQuery } from '@vue-storefront/core';
 import type { FacetData, SearchInputParams, Context } from '../../types';
 
 // TODO: Later refactor this to useSearch instead. useFacet will be deprecated soon
 const getFacet = async (context: Context, params: SearchInputParams, customQuery?: CustomQuery): Promise<ApolloQueryResult<FacetData>> => {
-  const defaultVariables = {
+  const searchVariables = {
     ...params
   };
 
   const { search } = context.extendQuery(customQuery,
-    { search: { query: defaultQuery, variables: defaultVariables } }
+    { search: { query: searchQuery, variables: searchVariables } }
   );
 
   const request = await context.client.query<FacetData>({
